Use cg-django-uaa for cloud.gov UAA auth.
