+++ .travis.yml
@@ -18,6 +18,7 @@ before_install:
 - node --version
 install:
 - pip install -r requirements-dev.txt
+- pip install -r TEMP_requirements.txt  # TODO remove this before final merge
 - npm install -g gulp-cli
 - npm install
 script:
