--- fake_uaa_provider/templates/fake_uaa_provider.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-{% load static from staticfiles %}
-<style>
-#logo {
-  width: 500px;
-}
-
-body {
-  font-family: "Source Sans Pro", Helvetica, Arial, sans-serif;
-  font-weight: 300;
-  max-width: 640px;
-  margin: 32px auto;
-  text-align: center;
-}
-
-a {
-  color: inherit;
-}
-
-footer {
-  padding-top: 3em;
-  color: gray;
-}
-
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0,0,0,0);
-  border: 0;
-}
-
-.submit-wrapper {
-  margin-top: 1em;
-}
-
-button[type="submit"] {
-  font-family: inherit;
-  font-weight: bold;
-  font-size: 1em;
-  padding: 16px;
-  border: none;
-  border-radius: 4px;
-  color: white;
-  background-color: #1F2C38;
-}
-
-input {
-  padding: 16px;
-  font-family: inherit;
-  font-size: 1em;
-  width: 20em;
-  border: 1px solid lightgray;
-}
-</style>
-<title>Fake cloud.gov Login</title>
-<body>
-  <img src="{% static 'fake_uaa_provider/fake-cloudgov.svg' %}" id="logo">
-  <p>Welcome to your fake <abbr title="User Account and Authentication">UAA</abbr> provider!</p>
-  <p>
-    Just enter an email address below and we'll log you in.
-  </p>
-  <form>
-    <label for="email" class="sr-only">E-mail</label>
-    <input type="email" name="email" id="email" autofocus>
-    {% for key, value in query.items %}
-      <input type="hidden" name="{{ key }}" value="{{ value }}">
-    {% endfor %}
-    <div class="submit-wrapper">
-      <button type="submit">Log In</button>
-    </div>
-  </form>
-  <footer>
-    Note that a user account with the corresponding email address needs
-    to exist in CALC's database for the login to successfully complete.
-    For more details, see the
-    <a href="https://github.com/18F/calc#readme">README</a>.
-  </footer>
-</body>
