+++ hourglass/templates/uaa_client/login_error.html
@@ -5,7 +5,7 @@
   <div class="row card">
     <div class="card__content">
       <h1>Login Error</h1>
-      {% if error_code == 'invalid_code_or_nonexistent_user' %}
+      {% if error_code == 'authenticate_failed' %}
         <p>
           Either you don't have permission to log in to this system, or
           an error occurred when trying to log you in.
